+++ CHANGELOG.md
@@ -12,6 +12,7 @@ To publish a new release run `scripts/release` from the project directory.
 - Add `--paginate` flag to paginate the output of mdcat (see [GH-152]);
   with this flag mdcat sends it output to `less -R` or any alternative pager set in `$MDCAT_PAGER` or `$PAGER`.
   This flag also disables all terminal-specific formatting as the pager likely won't support it.
+- Paginate output by default if invoked as `mdless`, that is, if `mdcat` is hard-linked to `mdless`.
 
 ### Fixed
 - Render email autolinks (i.e. `<hello@example.com>`) as `mailto:` links.